// src/app/models/userDetails.ts
export class UserDetails {


        address: string ;
        dob: string;
        emailId : string;
        firstName : string;
        gender : string;
        gstpNumber: string;
        lastName: string;
        maritalStatus: string;
        mobile: string;
        pinCode: string;
        qulification: string;
        status: string;
        userId: string;
      
      
}