// src/app/models/addUser.ts
export class AddUser {


    userId: 0 ;
    firstName: string;
    lastName : string;
    mobile : string;
    emailId : string;
    dob: string;
    gender: string;
    maritalStatus: string;
    gstpNumber: string;
    address: string;
    pinCode: string;
    qulification: string;
    status: "InActive";
  
}